[Integration][Github] Rename external custom properties executor module.
