+++ integrations/github/github/actions/registry.py
@@ -2,7 +2,7 @@ from port_ocean.context.ocean import ocean
 from github.actions.dispatch_workflow_executor import (
     DispatchWorkflowExecutor,
 )
-from github.actions.update_repo_external_properties_executor import (
+from github.actions.update_repo_external_custom_properties_executor import (
     UpdateRepoExternalCustomPropertiesExecutor,
 )
